+++ packages/components/scripts/vendorism.js
@@ -8,9 +8,9 @@ import { optimizePathForWindows } from 'vendorism/src/scripts/helpers.js';
 
 const components = ['input']
 const otherIncludes = ['custom-elements-manifest.config.js', 'web-test-runner.config.js', '*prettier*', 'tsconfig.json', 'src/declaration.d.ts', 'src/shoelace-autoloader*', 'src/translations/de.ts'];
-const libraryPrefix = 'sds';
-const libraryName = 'sick';
-const shoelaceVersion = '2.6.0';
+const libraryPrefix = 'syn';
+const libraryName = 'synergy';
+const shoelaceVersion = '2.10.0';
 
 //Command line options
 const optionDefinitions = [