+++ packages/components/src/styles/form-control.styles.ts
@@ -19,48 +19,48 @@ export default css`
   /* Label */
   .form-control--has-label .form-control__label {
     display: inline-block;
-    color: var(--sds-input-label-color);
-    margin-bottom: var(--sds-spacing-3x-small);
+    color: var(--syn-input-label-color);
+    margin-bottom: var(--syn-spacing-3x-small);
   }
 
   .form-control--has-label.form-control--small .form-control__label {
-    font-size: var(--sds-input-label-font-size-small);
+    font-size: var(--syn-input-label-font-size-small);
   }
 
   .form-control--has-label.form-control--medium .form-control__label {
-    font-size: var(--sds-input-label-font-size-medium);
+    font-size: var(--syn-input-label-font-size-medium);
   }
 
   .form-control--has-label.form-control--large .form-control__label {
-    font-size: var(--sds-input-label-font-size-large);
+    font-size: var(--syn-input-label-font-size-large);
   }
 
   :host([required]) .form-control--has-label .form-control__label::after {
-    content: var(--sds-input-required-content);
-    margin-inline-start: var(--sds-input-required-content-offset);
-    color: var(--sds-input-required-content-color);
+    content: var(--syn-input-required-content);
+    margin-inline-start: var(--syn-input-required-content-offset);
+    color: var(--syn-input-required-content-color);
   }
 
   /* Help text */
   .form-control--has-help-text .form-control__help-text {
     display: block;
-    color: var(--sds-input-help-text-color);
-    margin-top: var(--sds-spacing-3x-small);
+    color: var(--syn-input-help-text-color);
+    margin-top: var(--syn-spacing-3x-small);
   }
 
   .form-control--has-help-text.form-control--small .form-control__help-text {
-    font-size: var(--sds-input-help-text-font-size-small);
+    font-size: var(--syn-input-help-text-font-size-small);
   }
 
   .form-control--has-help-text.form-control--medium .form-control__help-text {
-    font-size: var(--sds-input-help-text-font-size-medium);
+    font-size: var(--syn-input-help-text-font-size-medium);
   }
 
   .form-control--has-help-text.form-control--large .form-control__help-text {
-    font-size: var(--sds-input-help-text-font-size-large);
+    font-size: var(--syn-input-help-text-font-size-large);
   }
 
   .form-control--has-help-text.form-control--radio-group .form-control__help-text {
-    margin-top: var(--sds-spacing-2x-small);
+    margin-top: var(--syn-spacing-2x-small);
   }
 `;