+++ packages/components/src/synergy-autoloader.ts
@@ -21,13 +21,13 @@ const observer = new MutationObserver(mutations => {
  */
 export async function discover(root: Element | ShadowRoot) {
   const rootTagName = root instanceof Element ? root.tagName.toLowerCase() : '';
-  const rootIsCustomElement = rootTagName?.includes('-');
+  const rootIsSynergyElement = rootTagName?.startsWith('syn-');
   const tags = [...root.querySelectorAll(':not(:defined)')]
     .map(el => el.tagName.toLowerCase())
-    .filter(tag => tag.startsWith('sds-'));
+    .filter(tag => tag.startsWith('syn-'));
 
-  // If the root element is an undefined custom element, add it to the list
-  if (rootIsCustomElement && !customElements.get(rootTagName)) {
+  // If the root element is an undefined Synergy component, add it to the list
+  if (rootIsSynergyElement && !customElements.get(rootTagName)) {
     tags.push(rootTagName);
   }
 
@@ -41,13 +41,13 @@ export async function discover(root: Element | ShadowRoot) {
  * Registers an element by tag name.
  */
 function register(tagName: string): Promise<void> {
-  const tagWithoutPrefix = tagName.replace(/^sds-/i, '');
-  const path = getBasePath(`components/${tagWithoutPrefix}/${tagWithoutPrefix}.js`);
-
   // If the element is already defined, there's nothing more to do
   if (customElements.get(tagName)) {
     return Promise.resolve();
   }
+
+  const tagWithoutPrefix = tagName.replace(/^syn-/i, '');
+  const path = getBasePath(`components/${tagWithoutPrefix}/${tagWithoutPrefix}.js`);
 
   // Register it
   return new Promise((resolve, reject) => {