+++ packages/components/src/synergy.ts
@@ -1,0 +1,12 @@
+// Components
+export { default as SynInput } from './components/input/input.js';
+/* plop:component */
+
+// Utilities
+// export * from './utilities/animation.js';
+// export * from './utilities/base-path.js';
+// export * from './utilities/icon-library.js';
+// export * from './utilities/form.js';
+
+// Events
+// export * from './events/events.js';